Move Tree View implementation to tree-view folder
