+++ src/components/tree-view/tree-x-list-item/tree-x-list-item.tsx
@@ -17,7 +17,7 @@ import {
   TreeXListItemNewCaption,
   TreeXListItemSelectedInfo
 } from "../tree-x/types";
-import { mouseEventModifierKey } from "../common/helpers";
+import { mouseEventModifierKey } from "../../common/helpers";
 
 // Drag and drop
 export type DragState = "enter" | "none" | "start";
