+++ src/components/tree-view/tree-x-list/readme.md
@@ -9,7 +9,7 @@
 
 ### Used by
 
- - [ch-test-tree-x](../test)
+ - [ch-test-tree-x](../../test)
 
 ### Graph
 ```mermaid
